fix: small fixes (not done)
