+++ yosoi/core/AGENTS.md
@@ -11,4 +11,4 @@ Contains the primary business logic and operational pipeline for Yosoi: network 
 5. **LLM Boundaries**: `pydantic_ai` and LLM interactions are strictly confined to the `discovery/` subdirectory.
 6. **Configuration**: `configs.py` holds `YosoiConfig`, `DebugConfig`, `TelemetryConfig`, and provider resolution logic.
 7. **Task Queue**: `tasks.py` holds taskiq broker setup, `process_url_task`, `enqueue_urls`, and concurrency helpers.
-8. **Pipeline Structure**: `Pipeline` is composed from focused mixin modules — `pipeline.py` (spine), `pipeline_cache.py` (cached replay), `pipeline_extraction.py` (fetch/clean/extract), `pipeline_discovery.py` (AI discovery/MCP/JS), `pipeline_crawler.py` (CAS-52 crawl helpers), `pipeline_utils.py` (stateless helpers). `_pipeline_table.py` holds the shared concurrent-progress table builder.
+8. **Pipeline Structure**: `Pipeline` is composed from focused mixin modules — `pipeline.py` (spine), `pipeline_cache.py` (cached replay), `pipeline_extraction.py` (fetch/clean/extract), `pipeline_discovery.py` (AI discovery/MCP/JS), `pipeline_crawler.py` (CAS-52 crawl helpers), `pipeline_utils.py` (stateless helpers). `_build_concurrent_table` is defined directly in `pipeline.py`.
